refactor(component): move loading to parent folder
